Update app routing
The routing in the application has changed. This is because the result
of 'ng build --prod' broke the app's routing.

It seems like this had something to do with the routes being created
dynamically in the source code? The compilation steps that uglify the
code broke the code-generated routes?

From the user perspective the only change is that the /pipe_inputs
end-point is now the /docs/pipe_inputs (analogous changes for all the
other documentation routes).

The Implementation is more different however. The routes, defined in
app-routing.module.ts, now use the templating syntax; '/docs/:repoName',
repo.service.ts now exposes a getRepoByName method;
repo-iframe.component.ts now calls the new RepoService method to get
render the appropriate route based in the sting obtained from the route
template; and lastly the sidenav routerLink arguments have changed to
reflect the changes to the route endpoints.